+++ src/main/java/ruslan/simakov/taxi/repository/CarRepository.java
@@ -4,8 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 import ruslan.simakov.taxi.model.Car;
 
-import java.util.List;
-
 @Repository
-public interface CarRepository extends CrudRepository<Car, Integer> {
+public interface CarRepository extends CrudRepository<Car, Long> {
 }
